requirements.txt
asgiref==3.6.0
attrs==22.2.0
common==0.1.2
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
psycopg2-binary==2.9.5
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
factory-boy==3.2.1
Faker==15.3.4
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
